+++ scripts/templates.mjs
@@ -11,6 +11,8 @@ async function preloadHandlebarsTemplates() {
 		// Item Sheet Partials
 		getModulePath("templates/items/details/details-equipment.hbs"),
 		getModulePath("templates/items/details/details-maneuver.hbs"),
+		getModulePath("templates/items/description-maneuver.hbs"),
+		getModulePath("templates/items/parts/maneuver-block.hbs"),
 		getModulePath("templates/items/chassis-panel.hbs"),
 		getModulePath("templates/items/chassis-install-browser.hbs"),
 		getModulePath("templates/apps/augmentations-manager.hbs"),